+++ changelog.md
@@ -5,6 +5,7 @@
 * LLT-3423: Add identifier to `Node` and `ExitNode`, and telio function to supply an identifier when connecting to an exit node
 * LLT-3404: Add github PR template
 * LLT-3468: Separate Heartbeat collection from QoS so heartbeat collection can be alone enabled.
+* LLT-3502: Refactor telio-nurse. Fix meshnet members list, conectivity matrix, external links data.
 
 ### v4.0.0
 ### **Bruschetta**